Fixed scaling bug.
If you apply SKTScaleEffect to a node with a physics body, then on iOS 7.1 and up this also scales the physics body, making collision detection unpredictable (it worked fine on iOS 7.0).

It's better to create a new parent node and give that the physics body, add the sprite/shape node as a child, and apply the SKTScaleEffect only to the child node.